Update ITPParser.py
As observed in https://github.com/BradyAJohnston/MolecularNodes/issues/944, if a GROMACS TOP file is provided as the input topology file, MDAnalysis is unable to guess it to be a GROMACS TOP file and always assumes it as an Amber PRMTOP file, causing the parser to fail. I tested making this change locally and it seems to solve the problem.